сущности добавлены в базу данных

+++ src/main/java/ru/naumen/javakids/model/Lecture.java
@@ -1,7 +1,7 @@
 package ru.naumen.javakids.model;
 
 public interface Lecture {
-    int getId();
+    Long getId();
     String getTopic();
     String getContent();
     Status getStatus();
